ExDM Add documentation to the selected parts of the code #369
- working on FunctionalProgramming
