+++ ExDataManagement/FunctionalProgramming/FunctionalProgramming.UnitTest/AnonymousFunctionsUnitTest.cs
@@ -1,6 +1,6 @@
 ﻿//____________________________________________________________________________________________________________________________________
 //
-//  Copyright (C) 2023, Mariusz Postol LODZ POLAND.
+//  Copyright (C) 2024, Mariusz Postol LODZ POLAND.
 //
 //  To be in touch join the community by pressing the `Watch` button and get started commenting using the discussion panel at
 //
@@ -19,11 +19,7 @@ namespace TP.FunctionalProgramming
   [TestClass]
   public class AnonymousFunctionsUnitTest
   {
-    [TestMethod]
-    public void StringIsLongPredicateTest()
-    {
-      Assert.IsTrue(AnonymousFunctions.StringIsLongPredicate("g5F|z*tC&yKJU$"));
-    }
+
 
     [TestMethod]
     public void NamedMethodCallBackTest()
