+++ ExDataManagement/FunctionalProgramming/FunctionalProgramming/FunctionalProgramming.csproj
@@ -44,12 +44,15 @@
     <Compile Include="AnonymousFunctions.cs" />
     <Compile Include="DelegateExample.cs" />
     <Compile Include="ExtensionMethods.cs" />
+    <Compile Include="FunctionalProgramming.cs" />
     <Compile Include="GlobalSuppressions.cs" />
-    <Compile Include="IMyInterface.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="README.AnonymousFunctions.md" />
+    <None Include="README.ExtensionMethods.md" />
     <None Include="Readme.md" />
+    <None Include="READMEDelegateEvents.md" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
