--- ExDataManagement/FunctionalProgramming/FunctionalProgramming/IMyInterface.cs
@@ -1,23 +1,0 @@
-﻿//____________________________________________________________________________________________________________________________________
-//
-//  Copyright (C) 2023, Mariusz Postol LODZ POLAND.
-//
-//  To be in touch join the community by pressing the `Watch` button and get started commenting using the discussion panel at
-//
-//  https://github.com/mpostol/TP/discussions/182
-//
-//  by introducing yourself and telling us what you do with this community.
-//_____________________________________________________________________________________________________________________________________
-namespace TP.FunctionalProgramming
-{
-  /// <summary>
-  /// Define an example interface named IMyInterface.
-  /// </summary>
-  public interface IMyInterface
-  {
-    /// <summary>
-    /// Any class that implements IMyInterface must implement the method that matches the following signature.
-    /// </summary>
-    void MyInterfaceMethod();
-  }
-}
